interp sel indexed passing tests
